deployment_package/requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
python-multipart
langchain
langgraph
langchain-openai
pydantic
httpx
jinja2
python-dotenv
structlog
resend
pytest
pytest-asyncio
respx
